requirements.txt
numpy
pandas
langchain
InstructorEmbedding
sentence-transformers
faiss-cpu
python-dotenv
google-generativeai
sentence-transformers
InstructorEmbedding